requirements.txt
attrs==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
coverage==7.0.0
crossplane==0.5.8
dominate==2.7.0
etcd3==0.12.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
grpcio==1.44.0
idna==3.3
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
protobuf==3.19.4
PyMySQL==1.0.2
pytest==7.2.0
pytest-cov==4.0.0
PyYAML==6.0
requests==2.27.1
six==1.12.0
tenacity==8.0.1
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.2.2
zipp==3.11.0
